# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation POS API",
        "description": "This API documentation outlines the features and endpoints of a Laravel project designed to master RESTful API development using Sanctum. It provides comprehensive documentation for secure user authentication, token-based sessions, and role-based access control. Developed as a learning tool, this project demonstrates best practices in Laravel API design, including resourceful routing, middleware application, and data validation techniques. It serves as a practical guide for developers aiming to implement modern API solutions with Laravel and Sanctum",
        "contact": {
            "email": "pichoendoo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/items/{item}/stocks": {
            "post": {
                "tags": [
                    "ItemStocks"
                ],
                "summary": "Create item stock",
                "description": "Create a new item stock",
                "operationId": "storeItemStock",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreItemStockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item stock created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemStock"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item stock created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create item stock. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/item-stocks/{itemStock}": {
            "get": {
                "tags": [
                    "ItemStocks"
                ],
                "summary": "Get item stock",
                "description": "Fetch a specific item stock by ID",
                "operationId": "showItemStock",
                "parameters": [
                    {
                        "name": "itemStock",
                        "in": "path",
                        "description": "Item Stock ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item stock fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemStock"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item stock fetched successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch item stock. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ItemStocks"
                ],
                "summary": "Update item stock",
                "description": "Update an existing item stock",
                "operationId": "updateItemStock",
                "parameters": [
                    {
                        "name": "itemStock",
                        "in": "path",
                        "description": "Item Stock ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateItemStockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item stock updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemStock"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item stock updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update item stock. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ItemStocks"
                ],
                "summary": "Delete item stock",
                "description": "Delete a specific item stock by ID",
                "operationId": "deleteItemStock",
                "parameters": [
                    {
                        "name": "itemStock",
                        "in": "path",
                        "description": "Item Stock ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item stock deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item stock deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete item stock. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreItemStockRequest": {
                "title": "StoreItemStockRequest",
                "description": "Store Item Stock Request",
                "required": [
                    "cogs",
                    "qty"
                ],
                "properties": {
                    "cogs": {
                        "type": "string",
                        "example": "Cost of goods sold"
                    },
                    "qty": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "UpdateItemStockRequest": {
                "title": "UpdateItemStockRequest",
                "description": "Update Item Stock Request Schema",
                "required": [
                    "item_id",
                    "cogs",
                    "qty"
                ],
                "properties": {
                    "item_id": {
                        "description": "Item ID (required)",
                        "type": "integer"
                    },
                    "cogs": {
                        "description": "Cost of Goods Sold (required)",
                        "type": "string"
                    },
                    "qty": {
                        "description": "Quantity (required)",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "ItemStock": {
                "title": "ItemStock",
                "required": [
                    "item_id",
                    "cogs",
                    "qty"
                ]
            },
...
        }
    },
...
}